+++ join_impl/src/chain/expr/initial_expr.rs
@@ -16,7 +16,7 @@ pub struct InitialExpr(pub [Expr; 1]);
 
 impl ToTokens for InitialExpr {
     fn to_tokens(&self, output: &mut TokenStream) {
-        let expr = self.get_inner().unwrap();
+        let expr = self.get_inner_exprs().unwrap();
         let tokens = quote! { #( #expr )* };
         output.extend(tokens);
     }
@@ -29,11 +29,11 @@ impl ToTokens for InitialExpr {
 }
 
 impl InnerExpr for InitialExpr {
-    fn get_inner(&self) -> Option<&[Expr]> {
+    fn get_inner_exprs(&self) -> Option<&[Expr]> {
         Some(&self.0)
     }
 
-    fn replace_inner(self, exprs: &[Expr]) -> Option<Self> {
+    fn replace_inner_exprs(self, exprs: &[Expr]) -> Option<Self> {
         exprs.last().cloned().map(|expr| Self([expr]))
     }
 }
@@ -52,7 +52,7 @@ mod tests {
         let expr: Expr = parse_quote! { |v| v + 1 };
 
         assert_eq!(
-            InitialExpr([expr.clone()]).get_inner().clone(),
+            InitialExpr([expr.clone()]).get_inner_exprs().clone(),
             Some(&[expr][..])
         );
     }
@@ -64,9 +64,9 @@ mod tests {
 
         assert_eq!(
             InitialExpr([expr])
-                .replace_inner(&[replace_inner.clone()][..])
+                .replace_inner_exprs(&[replace_inner.clone()][..])
                 .unwrap()
-                .get_inner()
+                .get_inner_exprs()
                 .clone(),
             Some(&[replace_inner][..])
         );